Header
Navbar och header färdigställt. Organiserat filerna
