+++ frontend/src/scenes/login.jsx
@@ -1,0 +1,10 @@
+import { Box } from "@mui/material";
+
+function LogIn() {
+
+    return(
+        <Box>
+            
+        </Box>
+    )
+}
